Move acceptance assertions to TCK
Rename test and remove assertions on result.
